backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-tenants>=3.5.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=5.0.0
django-cors-headers>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
reportlab>=4.0.0
channels>=4.0.0
daphne>=4.0.0
django-elasticsearch-dsl>=7.3.0
elasticsearch-dsl>=7.4.0
